api #1 and client #6: update map | update entities api: - update date in event, historical figure entities - update coordinates in landmark entity - add coordinates to event entity client: - add conection to data base - dynamic rendering events and landmarks on map - add interfaces for events and landmarks - update marker type

+++ apps/api/src/users/users.service.ts
@@ -1,10 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { UserRole } from './user-role.enum';
+import { UserRole } from '../auth/enums/role.enum';
 
 @Injectable()
 export class UsersService {
@@ -29,22 +29,40 @@ export class UsersService {
     return this.userRepository.find();
   }
 
-  async findOne(username: string): Promise<User | undefined> {
-    return this.userRepository.findOneBy({ username });
+  async findOne(criteria: {
+    username?: string;
+    email?: string;
+  }): Promise<User | undefined> {
+    const { username, email } = criteria;
+
+    if (username) {
+      return this.userRepository.findOneBy({ username });
+    }
+    if (email) {
+      return this.userRepository.findOneBy({ email });
+    }
+
+    return undefined;
   }
 
-  update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
-    const user: User = new User();
-    user.username = updateUserDto.username;
-    user.email = updateUserDto.email;
-    user.password = updateUserDto.password;
-    user.role = updateUserDto.role;
-    user.banStatus = updateUserDto.banStatus;
-    user.id = id;
-    return this.userRepository.save(user);
+  update(id: number, updateUserDto: UpdateUserDto, req: any): Promise<User> {
+    if (req.user.role === UserRole.Admin || req.user.userId === id) {
+      const user: User = new User();
+      user.username = updateUserDto.username;
+      user.email = updateUserDto.email;
+      user.password = updateUserDto.password;
+      user.role = updateUserDto.role;
+      user.banStatus = updateUserDto.banStatus;
+      user.id = id;
+      return this.userRepository.save(user);
+    }
+    throw new ForbiddenException();
   }
 
-  remove(id: number): Promise<{ affected?: number }> {
-    return this.userRepository.delete(id);
+  remove(id: number, req: any): Promise<{ affected?: number }> {
+    if (req.user.role === UserRole.Admin || req.user.userId === id) {
+      return this.userRepository.delete(id);
+    }
+    throw new ForbiddenException();
   }
 }
